Refactor
- Use self type annotation instead of extending
- Aggregate OIDC context to a case class
- Use companion object instead of dedicated utility class
